Modify content provider access.
Modified the permissions for the content provider access. Now it has explicit READ and WRITE permissions, that are required, at the moment to be mine too. This can be changed in the future but not right now. It is already prepared to be changed. The strings are defined on the strings.xml file